backend/pyproject.toml
[project]
name = "qr-checkin-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4",
  "pyjwt>=2.8",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov>=5", "httpx>=0.27"]

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing:skip-covered --cov-report=html --cov-fail-under=80"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
